# _Reversa_MasterSaaS/openapi/mastersaas-api.yaml
openapi: "3.1.0"
info:
  title: MasterSaaS API
  version: "1.0.0"
  description: |
    Especificação OpenAPI gerada pelo Reversa v1.2.14 a partir de engenharia reversa do frontend.
    O sistema atual opera em mock in-memory — esta spec define o contrato de backend a implementar.
    Autenticação delegada ao Supabase Auth; endpoints usam Bearer JWT emitido pelo Supabase.
servers:
  - url: https://api.mastersaas.com/v1
    description: Produção
  - url: http://localhost:3000/v1
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido pelo Supabase Auth (anon key ou service_role key)
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Recurso não encontrado"
...
    CurrencyCode:
      type: string
      enum: [BRL, USD, EUR, ARS, CLP, COP, MXN, PEN]
    ProductPrices:
      type: object
      additionalProperties:
        type: number
      description: "Mapa CurrencyCode → preço. Ex: { BRL: 197.00, USD: 39.00 }"
    Product:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          example: "evolia-pro"
        name:
          type: string
        description:
          type: string
          nullable: true
        prices:
          $ref: "#/components/schemas/ProductPrices"
        commissionRate:
          type: number
          minimum: 0
          maximum: 100
          description: "Taxa de comissão base em %"
        commissionDuration:
          type: string
          enum: [Lifetime, "12 months", "6 months", "3 months", Custom]
        active:
          type: boolean
          default: true
        productUrl:
          type: string
          format: uri
          description: "URL de checkout do produto"
        productCode:
          type: string
          description: "Código interno do produto para montagem de links de afiliado"
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Lista produtos do catálogo
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
          description: "Filtrar apenas produtos ativos"
        - name: currency
          in: query
          schema:
            $ref: "#/components/schemas/CurrencyCode"
          description: "Currency preferida para resolução de preço"
      responses:
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
              example:
                data:
                  - slug: "evolia-pro"
                    name: "Evolia Pro"
                    prices:
                      BRL: 197.00
                      USD: 39.00
                    commissionRate: 30
                    commissionDuration: "Lifetime"
                    active: true

    post:
      tags: [Products]
      summary: Cria ou atualiza produto (upsert por slug)
      security:
        - bearerAuth: []
      description: |
        Equivalente ao `upsertProduct`. Admin only.
        Se `slug` existir, atualiza. Caso contrário, cria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: Produto criado ou atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Sem permissão (não é admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...